iox-#381 Move every class into its own cpp/hpp
