+++ iceoryx_posh/include/iceoryx_posh/roudi/iceoryx_roudi_components.hpp
@@ -16,7 +16,6 @@
 
 #include "iceoryx_posh/internal/roudi/port_manager.hpp"
 #include "iceoryx_posh/internal/roudi/roudi_lock.hpp"
-#include "iceoryx_posh/internal/runtime/message_queue_interface.hpp"
 #include "iceoryx_posh/roudi/memory/iceoryx_roudi_memory_manager.hpp"
 #include "iceoryx_utils/cxx/expected.hpp"
 #include "iceoryx_utils/cxx/generic_raii.hpp"
