+++ iceoryx_posh/source/roudi/iceoryx_roudi_components.cpp
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include "iceoryx_posh/roudi/iceoryx_roudi_components.hpp"
+#include "iceoryx_posh/internal/runtime/ipc_base.hpp"
 
 namespace iox
 {
