+++ iceoryx_posh/test/moduletests/test_mq_interface.cpp
@@ -21,8 +21,8 @@
 #define private public
 #define protected public
 
-
-#include "iceoryx_posh/internal/runtime/message_queue_interface.hpp"
+#include "iceoryx_posh/internal/runtime/ipc_interface_user.hpp"
+#include "iceoryx_posh/internal/runtime/ipc_interface_creator.hpp"
 #include "iceoryx_posh/internal/runtime/message_queue_message.hpp"
 #include "iceoryx_utils/internal/posix_wrapper/ipc_channel.hpp"
